[Fix] Article ID is a long
https://github.com/cloudbees/zendesk-java-client/blob/master/src/main/java/org/zendesk/client/v2/model/hc/Article.java#L11
https://developer.zendesk.com/rest_api/docs/help_center/articles#json-format (marked as int but they aren't  - it is the same for others objects)

+++ src/main/java/org/zendesk/client/v2/Zendesk.java
@@ -1528,7 +1528,7 @@ public class Zendesk implements Closeable {
                 handleList(Article.class, "articles")));
     }
 
-    public Article getArticle(int id) {
+    public Article getArticle(long id) {
         return complete(submit(req("GET", tmpl("/help_center/articles/{id}.json").set("id", id)),
                 handle(Article.class, "article")));
     }
